Refactor duplicate image upload logic in productController
- Created `backend/utlis/imageUpload.js` utility function `uploadImage` to handle both buffer and string uploads to Cloudinary.
- Refactored `backend/controllers/productController.js` to use `uploadImage`, removing duplicate `upload_stream` and `upload` logic.
- Updated `backend/tests/productController.test.js` to unskip and fix tests by mocking `catchAsyncErrors`.
- Verified changes with existing tests `backend/tests/integration/productController_createProduct_multipart.test.js` and `backend/tests/productController.test.js`.
